requirements.txt
python-box~=6.0
shapely~=1.8
numpy~=1.22
numpy-quaternion~=2022.4
# 1.3.3 has a small bug for python3.10; fixed in live, waiting for proper release
p_tqdm @ https://github.com/swansonk14/p_tqdm.git
SimpleITK~=2.1.1
gcapi~=0.7
picai-prep @ https://github.com/DIAGNijmegen/picai_prep/archive/refs/tags/v1.1.1.zip